requirements.txt
requests==2.28.1
pandas
kucoin-python
